feat: search functionality dialog finished
