+++ src/layouts/BaseLayout.astro
@@ -3,6 +3,7 @@ import Navigation from "../components/Navigation.astro";
 import Footer from "../components/Footer.astro";
 import Sidebar from "../components/Sidebar.astro";
 import CartDrawer from "../components/CartDrawer.svelte";
+import SearchDialog from "$/components/SearchDialog.svelte"
 import AnnouncementBar from "../components/AnnouncementBar.astro";
 
 export interface Props {
@@ -46,5 +47,6 @@ const { title, description = defaultDesc, navigation } = Astro.props as Props;
       <slot />
     </main>
     <Footer />
+    <SearchDialog client:load/>
   </body>
 </html>
